pyproject.toml
[tool.poetry]
name = "test-bot"
version = "0.1.0"
description = ""
authors = ["Yevhenii Melnykov <yevhenii.melnykov@codeit.local>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
aiogram = "^3.17.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
